Fixed documentation, type, and usages of the OpenWorkspaceXMLDAO constructor.
Previously, there was no way to specify you didn't know the size of a stream,
but half the call sites of the constructor were passing in 0 to mean unknown.
The correct way to do this is now documented, there's a constant representing
unknown, AND the job size of the Monitorable is reported as unknown rather than
0 when the file size is unknown.

Additionally, many calls to the constructor were casting long to int (because
File.size() returns long) so I've adjusted the type to request long instead.

+++ src/main/java/ca/sqlpower/wabit/swingui/WabitSwingSessionContextImpl.java
@@ -1136,7 +1136,8 @@ public class WabitSwingSessionContextImpl implements WabitSwingSessionContext {
 					} catch (FileNotFoundException e) {
 						throw new RuntimeException(e);
 					}
-					OpenWorkspaceXMLDAO open = new OpenWorkspaceXMLDAO(delegateContext, input, (int) file.length());
+					OpenWorkspaceXMLDAO open =
+					    new OpenWorkspaceXMLDAO(delegateContext, input, file.length());
 					open.importWorkspaces(getActiveSession());
 					
 				}
